backend/pyproject.toml
[tool.poetry]
name = "noted"
version = "0.1.0"
description = "A Django Ninja API backend with authentication, profiles and images"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
django-ninja = "^1.0.1"
django-cors-headers = "^4.3.0"
Pillow = "^10.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
pydantic = "^2.4.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
email-validator = "^2.1.0"
channels = "^4.2.0"
channels-redis = "^4.2.1"
psycopg2 = "^2.9.10"
daphne = "^4.0.0"
uvicorn = "^0.32.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
